Fixed Board Gen, Wrote Tester, Simplified Lookups
Added GetSquare method with two overloads. Use that when doing gameGrid
lookups, as generating a new coordinate does not work for the key, and
otherwise you have to use a linq statement every time.

On a similar note, I rewrote the boardgen logic so it would work
properly, and change the piece class so that CurrentPosition actually
queries the gameBoard for the piece's own position.

Finally, I wrote a simple test app that proves that our pieces go where
they're supposed to, and can be looked up.