Huge Push
Player Class Added. Contains its king and a list of all its pieces for
ease of access. A player's pieces are the same piece objects that the
board holds, and are put there on piece generation.

Pieces also store the player they belong to as a Player object in
addition to its number, for ease of access.

All pieces now have ValidMoves and ValidThreat properties, which return
only the spaces they could move to without breaking the rules.